refactor: move filters code to component
As that's the only place it's used, and the store code gets included
everywhere.
