+++ src/components/search/SideFilters.vue
@@ -94,7 +94,7 @@
   import { mapState, mapGetters } from 'vuex';
   import { rangeToQueryParam, rangeFromQueryParam } from '@/plugins/europeana/search';
   import themes from '@/plugins/europeana/themes';
-  import { queryUpdatesForFilters } from '../../store/search';
+  import { defaultFacetNames } from '@/store/search';
   import SideFacetDropdown from './SideFacetDropdown';
 
   export default {
@@ -133,7 +133,6 @@
       ...mapGetters({
         facetNames: 'search/facetNames',
         filters: 'search/filters',
-        queryUpdatesForFacetChanges: 'search/queryUpdatesForFacetChanges',
         collection: 'search/collection'
       }),
       resetButtonDisabled() {
@@ -241,6 +240,57 @@
 
         this.rerouteSearch(this.queryUpdatesForFacetChanges({ [name]: selected }));
       },
+      queryUpdatesForFacetChanges(selected = {}) {
+        const filters = Object.assign({}, this.filters);
+
+        for (const name in selected) {
+          filters[name] = selected[name];
+        }
+
+        // Remove collection-specific filters when collection is changed
+        if (Object.prototype.hasOwnProperty.call(selected, 'collection') || !this.collection) {
+          for (const name in filters) {
+            if (name !== 'collection' && !defaultFacetNames.includes(name) && this.resettableFilters.includes(name)) {
+              filters[name] = [];
+            }
+          }
+        }
+
+        // Remove filters incompatible with change of collection filter
+        if (Object.prototype.hasOwnProperty.call(selected, 'collection') && Object.prototype.hasOwnProperty.call(filters, 'contentTier')) {
+          filters['contentTier'] = [];
+        }
+
+        return this.queryUpdatesForFilters(filters);
+      },
+      queryUpdatesForFilters(filters) {
+        const queryUpdates = {
+          qf: [],
+          page: 1
+        };
+
+        for (const name in filters) {
+          switch (name) {
+          case 'REUSABILITY':
+            // `reusability` has its own API parameter and can not be queried in `qf`
+            queryUpdates.reusability = filters[name].length > 0 ? filters[name].join(',') : null;
+            break;
+          case 'api':
+            // `api` is an option to /plugins/europeana/search/search()
+            queryUpdates.api = filters[name];
+            break;
+          default:
+            // Everything else goes in `qf`
+            queryUpdates.qf = queryUpdates.qf.concat(this.queryUpdatesForFilter(name, filters[name]));
+          }
+        }
+        return queryUpdates;
+      },
+      queryUpdatesForFilter(name, values) {
+        return [].concat(values)
+          .filter((value) => (value !== undefined) && (value !== null))
+          .map((value) => `${name}:${value}`);
+      },
       rerouteSearch(queryUpdates) {
         const query = this.updateCurrentSearchQuery(queryUpdates);
         this.$goto(this.$path({ ...this.route, ...{ query } }));
@@ -281,7 +331,7 @@
           filters[filterName] = [];
         }
         this.$store.commit('search/clearResettableFilters');
-        return this.rerouteSearch(queryUpdatesForFilters(filters));
+        return this.rerouteSearch(this.queryUpdatesForFilters(filters));
       },
       isFilteredByDropdowns() {
         return this.$store.getters['search/hasResettableFilters'];
