+++ tests/unit/store/search.spec.js
@@ -99,128 +99,6 @@ describe('store/search', () => {
       });
     });
 
-    describe('queryUpdatesForFacetChanges', () => {
-      const state = {
-        resettableFilters: []
-      };
-      const getters = {};
-
-      describe('when facet is REUSABILITY', () => {
-        describe('with values selected', () => {
-          const selected = { 'REUSABILITY': ['open', 'permission'] };
-          it('sets `reusability` to values joined with ","', () => {
-            const updates = store.getters.queryUpdatesForFacetChanges(state, getters)(selected);
-            expect(updates.reusability).toBe('open,permission');
-          });
-        });
-
-        describe('with no values selected', () => {
-          it('sets `reusability` to `null`', () => {
-            const updates = store.getters.queryUpdatesForFacetChanges(state, getters)();
-            expect(updates).toEqual({ qf: [], page: 1 });
-          });
-        });
-      });
-
-      describe('for default facets from search plugin supporting quotes', () => {
-        it('includes fielded and quoted queries for each value in `qf`', () => {
-          const selected = { 'TYPE': ['"IMAGE"', '"SOUND"'] };
-          const updates = store.getters.queryUpdatesForFacetChanges(state, getters)(selected);
-          expect(updates.qf).toContain('TYPE:"IMAGE"');
-          expect(updates.qf).toContain('TYPE:"SOUND"');
-        });
-      });
-
-      describe('for default facets from search plugin not supporting quotes', () => {
-        it('includes fielded but unquoted queries for each value in `qf`', () => {
-          const selected = { 'MIME_TYPE': ['application/pdf'] };
-          const updates = store.getters.queryUpdatesForFacetChanges(state, getters)(selected);
-          expect(updates.qf).toContain('MIME_TYPE:application/pdf');
-        });
-      });
-
-      describe('for any other facets', () => {
-        it('includes fielded but unquoted queries for each value in `qf`', () => {
-          const selected = { 'contentTier': ['4'] };
-          const updates = store.getters.queryUpdatesForFacetChanges(state, getters)(selected);
-          expect(updates.qf).toContain('contentTier:4');
-        });
-      });
-
-      describe('in a collection having custom filters', () => {
-        const state = {
-          userParams: {
-            qf: ['proxy_dcterms_issued:1900-01-01']
-          },
-          resettableFilters: ['proxy_dcterms_issued']
-        };
-        const getters = {
-          collection: () => 'newspaper'
-        };
-
-        it('applies them', () => {
-          const selected = { api: 'metadata', 'proxy_dcterms_issued': ['1900-01-02'] };
-
-          const updates = store.getters.queryUpdatesForFacetChanges(state, getters)(selected);
-
-          expect(updates.qf).toContain('proxy_dcterms_issued:1900-01-02');
-          expect(updates.api).toBe('metadata');
-        });
-      });
-
-      describe('with collection-specific facets already selected', () => {
-        const state = {
-          resettableFilters: ['collection', 'CREATOR', 'TYPE']
-        };
-        const getters = {
-          filters: {
-            'CREATOR': ['"Missoni (Designer)"'],
-            'TYPE': ['"IMAGE"'],
-            'contentTier': ['*']
-          },
-          collection: 'fashion'
-        };
-
-        describe('when collection is changed', () => {
-          const selected = { 'collection': 'art' };
-
-          it('removes collection-specific facet filters', () => {
-            const updates = store.getters.queryUpdatesForFacetChanges(state, getters)(selected);
-
-            expect(updates.qf).not.toContain('CREATOR:"Missoni (Designer)"');
-          });
-
-          it('preserves generic facet filters', () => {
-            const updates = store.getters.queryUpdatesForFacetChanges(state, getters)(selected);
-
-            expect(updates.qf).toContain('TYPE:"IMAGE"');
-          });
-
-          it('removes tier filter', () => {
-            const updates = store.getters.queryUpdatesForFacetChanges(state, getters)(selected);
-
-            expect(updates.qf).not.toContain('contentTier:*');
-          });
-        });
-
-        describe('when collection is removed', () => {
-          const selected = { 'collection': null };
-
-          it('removes collection-specific facet filters', () => {
-            const updates = store.getters.queryUpdatesForFacetChanges(state, getters)(selected);
-
-            expect(updates.qf).not.toContain('CREATOR:"Missoni (Designer)"');
-          });
-
-          it('preserves generic facet filters', () => {
-            const updates = store.getters.queryUpdatesForFacetChanges(state, getters)(selected);
-
-            expect(updates.qf).toContain('TYPE:"IMAGE"');
-          });
-        });
-      });
-    });
-
     describe('hasCollectionSpecificSettings', () => {
       describe('when collection param is absent', () => {
         const collection = undefined;
